move n-best logic to analysis-common
